refactor: normalize source module names
