+++ src/core/parser/srecParser.ts
