+++ src/webview/memory/memoryData.ts
@@ -1,7 +1,7 @@
 // Webview state adapter for core memory helpers.
 
-import { buildMemoryRows, buildSegmentIndex, getByteAt } from '../core/memory';
-import { S, BPR } from './state';
+import { buildMemoryRows, buildSegmentIndex, getByteAt } from '../../core/memory';
+import { S, BPR } from '../state';
 
 export function getByte(addr: number): number | undefined {
     return getByteAt(S.parseResult, S.segmentIndex, S.edits, addr);
