+++ src/webview/renderRegistry.ts
