+++ src/webview/sidebar/sidebarTypes.ts
